+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - SDAP-393: Included `insitu` in ingress based on the value of `insituAPI.enabled` in `values.yaml`
 - SDAP-371: Renamed `/domssubset` endpoint to `/cdmssubset`
 - SDAP-396: Added saildrone insitu api to matchup
+- CDMS-151: Added script for regression tests.
 ### Changed
 ### Deprecated
 ### Removed